Move library functions out of verilog_modeling to "models.utils".
